+++ contracts/ecdsa/ZkOwnershipECDSA.sol
@@ -3,18 +3,16 @@ pragma solidity 0.8.10;
 
 import {Address} from "@openzeppelin/contracts/utils/Address.sol";
 import {ECDSA} from "@openzeppelin/contracts/utils/cryptography/ECDSA.sol";
-import {IEddsaMimcVerifier} from "./IEddsaMimcVerifier.sol";
-import "hardhat/console.sol";
+import {IECDSAVerifier} from "./IECDSAVerifier.sol";
 
 /// @title The fund action task executor
-contract ZkOwnershipEddsaMimc {
+contract ZkOwnershipECDSA {
     using Address for address;
     using ECDSA for bytes32;
 
     uint256 public nonce;
-    uint256 public pubkeyX;
-    uint256 public pubkeyY;
-    IEddsaMimcVerifier public immutable verifier;
+    uint256 public pubkeyHash;
+    IECDSAVerifier public immutable verifier;
 
     struct ProofsRelated {
         uint256[2] a;
@@ -22,14 +20,9 @@ contract ZkOwnershipEddsaMimc {
         uint256[2] c;
     }
 
-    constructor(
-        IEddsaMimcVerifier verifier_,
-        uint256 pubkeyX_,
-        uint256 pubkeyY_
-    ) {
+    constructor(IECDSAVerifier verifier_, uint256 pubkeyHash_) {
         verifier = verifier_;
-        pubkeyX = pubkeyX_;
-        pubkeyY = pubkeyY_;
+        pubkeyHash = pubkeyHash_;
     }
 
     // / @notice Task execution function, will charge execution fee first.
@@ -44,7 +37,7 @@ contract ZkOwnershipEddsaMimc {
         ProofsRelated memory proof = _unpackProof(proof_);
 
         // Get public input
-        uint256[6] memory input = _getPublicInput(to_, execData_);
+        uint256[5] memory input = _getPublicInput(to_, execData_);
 
         // Verify proof
         require(verifier.verifyProof(proof.a, proof.b, proof.c, input), "Verify proof fail.");
@@ -57,10 +50,9 @@ contract ZkOwnershipEddsaMimc {
 
     // @notice set pubkey hash. Update only by address(this)
     // @param pubkeyHash_ New zk proof of ownership hash.
-    function setPubkey(uint256 newPubkeyX_, uint256 newPubkeyY_) external payable {
+    function setPubkeyHash(uint256 pubkeyHash_) external payable {
         require(msg.sender == address(this), "Not permitted");
-        pubkeyX = newPubkeyX_;
-        pubkeyY = newPubkeyY_;
+        pubkeyHash = pubkeyHash_;
     }
 
     // ----------------
@@ -76,18 +68,16 @@ contract ZkOwnershipEddsaMimc {
             });
     }
 
-    function _getPublicInput(address to_, bytes calldata execData_) internal view returns (uint256[6] memory) {
-        bytes32 msgHash = keccak256(abi.encodePacked(nonce, to_, execData_));
+    function _getPublicInput(address to_, bytes calldata execData_) internal view returns (uint256[5] memory) {
+        bytes32 msgHash = keccak256(abi.encodePacked(nonce, to_, execData_)).toEthSignedMessageHash();
+        uint256[5] memory input;
         bytes16 halfL = bytes16(msgHash);
         bytes16 halfR = bytes16(uint128(uint256(msgHash)));
-
-        uint256[6] memory input;
-        input[0] = pubkeyX;
-        input[1] = pubkeyY;
-        input[2] = uint256(uint64(bytes8(uint64(uint128(halfR)))));
-        input[3] = uint256(uint64(bytes8(halfR)));
-        input[4] = uint256(uint64(bytes8(uint64(uint128(halfL)))));
-        input[5] = uint256(uint64(bytes8(halfL)));
+        input[0] = pubkeyHash;
+        input[1] = uint256(uint64(bytes8(uint64(uint128(halfR)))));
+        input[2] = uint256(uint64(bytes8(halfR)));
+        input[3] = uint256(uint64(bytes8(uint64(uint128(halfL)))));
+        input[4] = uint256(uint64(bytes8(halfL)));
         return input;
     }
 }