+++ scripts/build_verify.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
 set -e
 
-PHASE1=../circuits/pot21_final.ptau
+PHASE1=../pot/pot$2_final.ptau
 BUILD_DIR=../build
-CIRCUIT_NAME=ownership
+CIRCUIT_NAME=$1
+
 
 if [ -f "$PHASE1" ]; then
     echo "Found Phase 1 ptau file"
@@ -52,13 +53,32 @@ echo "DONE ($((end-start))s)"
 
 echo "** Exporting vkey"
 start=`date +%s`
-npx snarkjs zkey export verificationkey "$BUILD_DIR"/"$CIRCUIT_NAME".zkey "$BUILD_DIR"/vkey.json
+npx snarkjs zkey export verificationkey "$BUILD_DIR"/"$CIRCUIT_NAME".zkey "$BUILD_DIR"/"$CIRCUIT_NAME"_vkey.json
 end=`date +%s`
 echo "DONE ($((end-start))s)"
 
 # Generate verifier contract
 echo "** BUILD VERIFIER CONTRACT **"
 start=`date +%s`
-snarkjs zkey export solidityverifier "$BUILD_DIR"/"$CIRCUIT_NAME".zkey ../contracts/Verifier.sol
+snarkjs zkey export solidityverifier "$BUILD_DIR"/"$CIRCUIT_NAME".zkey ../contracts/$3Verifier.sol
+end=`date +%s`
+echo "DONE ($((end-start))s)"
+
+
+# Change Solidity compiler version and contract name
+echo "** MODIFT VERIFIER CONTRACT COMPLIER VERSION**"
+start=`date +%s`
+sed -i '' 's/0.6.11;/0.8.10;\n\n/' "../contracts/$3Verifier.sol"
+sed -i '' "s/contract Verifier {/contract $3Verifier {/" "../contracts/$3Verifier.sol"
+yarn prettify
+end=`date +%s`
+echo "DONE ($((end-start))s)"
+
+# Copy all the required files to the public directory
+echo "** COPY ALL REQUIRED FILES TO PUBLIC DIRECTORY**"
+start=`date +%s`
+cp "$BUILD_DIR"/"$CIRCUIT_NAME".zkey "../public"/"$CIRCUIT_NAME".zkey
+cp "$BUILD_DIR"/"$CIRCUIT_NAME"_vkey.json "../public"/"$CIRCUIT_NAME"_vkey.json
+cp "$BUILD_DIR/$1_js/$1.wasm" "../public/$1.wasm"
 end=`date +%s`
 echo "DONE ($((end-start))s)"