+++ CHANGELOG.md
@@ -6,4 +6,6 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- EDDSA MIMC ownership circuit and contracts
+- ECDSA ownership circuit and contracts
 - Initial commit.